Update README.md
in our website, charmchronicle, we have made this website on travel blog plateform to help people know about places and travel there. We will be working on it more, and add many feautures to it. Also, we put the link of a website from which people can book their tickets and visit places easily. Our mission is to help tourism of our country get promoted and people know abpout tourist places. 
home page: